requirements.txt
Pillow>=10.0
pygame>=2.5
python-dotenv>=1.0
bcrypt>=4.0
